refactor operator prompt for re-usability
