+++ app/packages/operators/src/OperatorPrompt.tsx
@@ -1,5 +1,4 @@
-import { Box, LinearProgress, Typography } from "@mui/material";
-import { ComponentType, useCallback } from "react";
+import { ComponentType, ReactElement } from "react";
 import { createPortal } from "react-dom";
 import { useRecoilValue } from "recoil";
 import OperatorIO from "./OperatorIO";
@@ -8,16 +7,14 @@ import {
   useOperatorPrompt,
   useShowOperatorIO,
 } from "./state";
-// todo: use plugin component
-import ErrorView from "../../core/src/plugins/SchemaIO/components/ErrorView";
 import {
   BaseStylesProvider,
   PaletteContentContainer,
 } from "./styled-components";
-import OperatorPalette, { OperatorPaletteProps } from "./OperatorPalette";
-import { formatValidationErrors, stringifyError } from "./utils";
-import { types } from ".";
+import OperatorPalette from "./OperatorPalette";
+import { getOperatorPromptConfigs } from "./utils";
 import { CUSTOM_PROMPTS } from "./constants";
+import OperatorPromptBody from "./components/OperatorPromptBody";
 
 export default function OperatorPrompt() {
   const show = useRecoilValue(showOperatorPromptSelector);
@@ -44,55 +41,24 @@ export function withOperatorPrompt(Component: ComponentType) {
 
 function ActualOperatorPrompt() {
   const operatorPrompt = useOperatorPrompt();
-
-  const showResultOrError =
-    operatorPrompt.hasResultOrError ||
-    operatorPrompt.executorError ||
-    operatorPrompt.resolveError;
-  const showPrompt = operatorPrompt.showPrompt;
-  const isExecuting = operatorPrompt.isExecuting;
-  const show = showResultOrError || showPrompt || isExecuting;
-  const customPrompt = operatorPrompt?.promptView;
-  const customPromptName = customPrompt?.name;
+  const {
+    title,
+    hasValidationErrors,
+    resolving,
+    pendingResolve,
+    validationErrorsStr,
+    show,
+    customPromptName,
+    ...otherConfigs
+  } = getOperatorPromptConfigs(operatorPrompt);
 
   if (CUSTOM_PROMPTS.includes(customPromptName) || !show) return null;
 
-  const paletteProps: OperatorPaletteProps = {
-    submitButtonText: "Execute",
-    submitButtonOptions: operatorPrompt.submitOptions.options,
-    submitButtonLoading: operatorPrompt.submitOptions.isLoading,
-    hasSubmitButtonOptions: operatorPrompt.submitOptions.hasOptions,
-    showWarning: operatorPrompt.submitOptions.showWarning,
-    warningMessage: operatorPrompt.submitOptions.warningMessage,
-    cancelButtonText: "Cancel",
-  };
-
-  if (customPromptName == "PromptView") {
-    const prompt = types.PromptView.fromJSON(customPrompt);
-    if (prompt.submitButtonLabel)
-      paletteProps.submitButtonText = prompt.submitButtonLabel;
-  }
-
-  if (operatorPrompt.showPrompt) {
-    paletteProps.onSubmit = operatorPrompt.onSubmit;
-    paletteProps.onCancel = operatorPrompt.cancel;
-  } else if (showResultOrError) {
-    paletteProps.onCancel = operatorPrompt.close;
-    paletteProps.cancelButtonText = "Close";
-  }
-
-  const title = getPromptTitle(operatorPrompt);
-  const hasValidationErrors = operatorPrompt.validationErrors?.length > 0;
-  const { resolving, pendingResolve } = operatorPrompt;
-  const validationErrorsStr = formatValidationErrors(
-    operatorPrompt.validationErrors
-  );
-
   return createPortal(
     <OperatorPalette
+      {...otherConfigs}
       title={title}
-      {...paletteProps}
-      onClose={paletteProps.onCancel || operatorPrompt.close}
+      onClose={otherConfigs.onCancel || operatorPrompt.close}
       submitOnControlEnter
       disableSubmit={hasValidationErrors || resolving || pendingResolve}
       disabledReason={
@@ -104,51 +70,12 @@ function ActualOperatorPrompt() {
       loading={resolving || pendingResolve}
     >
       <PaletteContentContainer>
-        {operatorPrompt.showPrompt && (
-          <Prompting operatorPrompt={operatorPrompt} />
-        )}
-        {operatorPrompt.isExecuting && <Executing />}
-        {showResultOrError && (
-          <ResultsOrError
-            operatorPrompt={operatorPrompt}
-            outputFields={operatorPrompt.outputFields}
-          />
-        )}
+        <OperatorPromptBody operatorPrompt={operatorPrompt} />
       </PaletteContentContainer>
     </OperatorPalette>,
     document.body
   );
 }
-
-function Executing() {
-  return (
-    <Box>
-      <LinearProgress />
-      <Typography sx={{ pt: 1, textAlign: "center" }}>Executing...</Typography>
-    </Box>
-  );
-}
-
-export function Prompting({ operatorPrompt }) {
-  const setFormState = useCallback((data) => {
-    const formData = data;
-    for (const field in formData) {
-      operatorPrompt.setFieldValue(field, formData[field]);
-    }
-  }, []);
-
-  return (
-    <Box component={"form"} p={2} onSubmit={operatorPrompt.onSubmit}>
-      <OperatorIO
-        schema={operatorPrompt.inputFields}
-        onChange={setFormState}
-        data={operatorPrompt.promptingOperator.params}
-        errors={operatorPrompt?.validationErrors || []}
-      />
-    </Box>
-  );
-}
-
 export function OperatorViewModal() {
   const io = useShowOperatorIO();
   if (!io.visible) return null;
@@ -164,38 +91,6 @@ export function OperatorViewModal() {
       </PaletteContentContainer>
     </OperatorPalette>,
     document.body
-  );
-}
-
-function ResultsOrError({ operatorPrompt, outputFields }) {
-  const executorError = operatorPrompt?.executorError;
-  const resolveError = operatorPrompt?.resolveError;
-  const error = resolveError || executorError;
-  if (!outputFields && !executorError && !resolveError) return null;
-  const { result } = operatorPrompt.executor;
-
-  return (
-    <Box p={2}>
-      {outputFields && (
-        <OperatorIO
-          type="output"
-          data={result}
-          schema={operatorPrompt.outputFields}
-          onChange={() => {}}
-        />
-      )}
-      {error && (
-        <ErrorView
-          schema={{ view: { detailed: true } }}
-          data={[
-            {
-              reason: "Error occurred during operator execution",
-              details: stringifyError(error),
-            },
-          ]}
-        />
-      )}
-    </Box>
   );
 }
