+++ app/packages/operators/src/components/OperatorPromptHeader.tsx
@@ -1,0 +1,10 @@
+import { ReactElement } from "react";
+import { BaseStylesProvider } from "../styled-components";
+
+export default function OperatorPromptHeader({
+  title,
+}: {
+  title: ReactElement;
+}) {
+  return <BaseStylesProvider>{title}</BaseStylesProvider>;
+}
